Working on the remove account function. The program is able to iterate through the CSV and output each line to the console. However, at this time I am unable to get just one specific column. The internet showed me how to get "everything left of the deliminter" but that only returns the amount, not the name of the account. ( I can also use line[0] but that returns the fitst character of the string, not the entire string in the column.
I could switch the column order at creation, but it would be helpful to access each column separtley.

I am th thinking that I will need to add the items to a vector so that I can listdisplay them and the user can select which account (by name) that they would like to remove.